4th commit

+++ src/main/java/com/foodrecipe/api/repository/IngredientsRepository.java
@@ -3,6 +3,6 @@ package com.foodrecipe.api.repository;
 import com.foodrecipe.api.entity.Ingredients;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngredientsRepository extends JpaRepository<Ingredients, Integer> {
+public interface IngredientsRepository extends JpaRepository<Ingredients, Long> {
 
 }
